fixed CI issues
