+++ app/src/main/java/com/example/streamtv/utils/Validator.kt
@@ -1,4 +1,4 @@
-package com.example.streamtv.Utils
+package com.example.streamtv.utils
 
 import android.util.Patterns
 import android.widget.EditText
