+++ app/src/main/res/layout/login.xml
@@ -4,7 +4,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Login.LoginFragment">
+        tools:context=".login.LoginFragment">
 
     <ImageView
             android:id="@+id/logo"
